(hds-1467) Fix and combine mixed hook tests
